+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - GH-40 Fix encoding/decoding variable length field - ([e7d7f87](e7d7f87106b13033ae4eb9d053e8ab4c25ca568d)) - GH-40 (https://github.com/eastern-oak/tjiftjaf/issues/40)
+- GH-45 Fix panic on `Frame.header()` - ([5373520](537352019d55264d7afe5afa0215d31908ce1584)) - GH-45 (https://github.com/eastern-oak/tjiftjaf/issues/45)
 
 ## [0.1.0] - 2025-11-02